requirements.txt
argcomplete==1.12.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.12.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
av==8.0.3
boto3==1.24.54
botocore==1.27.54
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
crc32c==2.2.post0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.15
django-allauth==0.51.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-heroku==0.3.1
django-resized==1.0.2
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.13.1
dnspython==2.1.0
docutils==0.19
draftjs-exporter==5.0.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
jmespath==1.0.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
msgpack==1.0.4
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pep8==1.7.1
Pillow==9.2.0
pipx==0.16.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyee==8.1.0
PyHamcrest==2.0.4
PyJWT==2.4.0
pylibsrtp==0.6.8
pynetstring==0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==4.0.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.11
userpath==1.7.0
virtualenv==16.4.3
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
